pyproject.toml
[project]
name = "ict-bank"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt==4.1.1",
    "fastapi==0.104.1",
    "pandas==2.1.3",
    "passlib==1.7.4",
    "psycopg2-binary==2.9.9",
    "pydantic==2.5.0",
    "python-multipart==0.0.6",
    "requests==2.31.0",
    "streamlit==1.28.1",
    "uvicorn==0.24.0",
]
